src/requirements.txt
asgiref==3.3.1
coverage==5.3.1
Django==3.1.5
django-cors-headers==3.6.0
djangorestframework==3.12.2
python-decouple==3.4
pytz==2020.5
sqlparse==0.4.1
